Remove implicit convert from int to half

+++ src/Native/src/kernels/stackvm/reference/lstm.cpp
@@ -145,8 +145,8 @@ result<void> lstm_impl(const T *input, const T *w_xc, const T *w_rc,
 
                 // tanh_ct = tanh(ct_o)
                 for (size_t o = 0; o < out_shape[3]; o++) {
-                    out_mul1[o + out_shape[3] * 3] = tanh(float(
-                        output_c_tmp[o + d * out_shape[2] * out_shape[3]]));
+                    out_mul1[o + out_shape[3] * 3] = std::tanh(
+                        output_c_tmp[o + d * out_shape[2] * out_shape[3]]);
                 }
 
                 // ht = ot * tanh_ct
